Added the `library_generate` module to `rusty-junctions-macro` crate
This allows the entire library to be generated to any given order,
simply by supplying the order to this macro.

Additionally, the `rusty-junctions-macro` crate type has been updated
to be `proc-macro`, this allows the crate root to have all of the
macro definitions and simply call the appreciate helper.
Which is much simpler than the nested crate architecture.