Add support for `Junction` naming
By default a name is generated for the `Junction` that is being using
a 128-bit UUID.
However, when the `<name> as Junction` is specified before the channel
definitions, the junction will by accessible in the current scope with
the given name.

This is essential for ensure the macro provides equivalent
functionality to the original syntax.